Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250519/사각형 채우기 3/rectangle-fill-3.cpp
@@ -7,7 +7,7 @@ int n;
 int main() {
     cin >> n;
 
-    int dp[1001] = {0};
+    long long dp[1001] = {0};
     dp[0] = 1;
     dp[1] = 2;
     dp[2] = 7;
